refactor: deduplicate mount hook types into MountManager
Move OnMountArgs, OnMountHook, and OnMountResult to mount-manager.ts as
the single source of truth. Add setContext() to MountManager so it
receives sandbox/workspace references, removing the wrapper closure in
Workspace. Type-only imports for Workspace avoid circular dependencies.
